requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
openai>=1.50.0
twilio>=9.0.0
deepgram-sdk>=3.4.0
elevenlabs>=1.0.0
resend>=2.0.0
chromadb>=0.5.0
pydantic>=2.0.0
httpx>=0.27.0
